[BOTF-81] Slit master sidebar into to components (#77)
## [BOTF-81] Description
Created sub-components to store the sections
